refactor: Split usecase to view and container
