+++ packages/example-usecase/src/NeverUpdate.js
@@ -1,0 +1,17 @@
+import React from 'react'
+
+class NeverUpdate extends React.Component {
+  props: {
+    children?: any
+  }
+
+  shouldComponentUpdate () {
+    return false
+  }
+
+  render () {
+    return <div>{this.props.children}</div>
+  }
+}
+
+export default NeverUpdate
